licensing concerns
Included alternate, similar license intended to serve as a catch-all against misuse of the original license against myself, the original author of this project (not to imply all assets are owned by myself solely; node_modules present are not authored by myself, and it is possible files within the directory structure of the repository are not mine. If I am considered a target by either lawyers or other legal representatives of any companies/agencies/groups/individuals who may be considering the contents of this project to be infringing upon their assumed rights, this license shall serve as the licensing terms, as I am not aware of any true illegality present within this project.